requirements.txt
requests
ics
python-dotenv
pytz